- fixed a few warnings

+++ include/aprilui/AnimatorFrameAnimation.h
@@ -27,9 +27,9 @@ namespace aprilui
 			hstr getImageBaseName() { return mImageBaseName; }
 			void setImageBaseName(chstr value) { mImageBaseName = value; }
 			int getFirstFrame() { return mFirstFrame; }
-			void setFirstFrame(float value) { mFirstFrame = value; }
+			void setFirstFrame(int value) { mFirstFrame = value; }
 			int getFrameCount() { return mFrameCount; }
-			void setFrameCount(float value) { mFrameCount = value; }
+			void setFrameCount(int value) { mFrameCount = value; }
 			void setProperty(chstr name, chstr value);
 			void notifyEvent(chstr name, void* params);
